Fix bug when command-line is too long
The +max-command-line-length+ was not accounting for a digest being present when writing.
Instead, write the digest as soon as we calculate it.
Not only does this work for the first command, but once we implement long-lived connections we can omit the digest on subsequent commands, so pjlink-get/set don't need the digest at all